refactor(core): rename resources and scopes query file name
rename plural filename
