+++ packages/core/src/queries/resource.ts
