+++ packages/core/src/queries/scope.ts
